Seperate the Adapter and Viewholder class to increase readability
